Distributed workers using Pub/Sub queues.
- Replace previous Kubernetes pod/job creation code. This didn't work
  well when trying run large scale analysis.

- Factor out analysis code into "analysis" package. Re-use this for the
  existing "analyze" binary and a new "worker" binary.
