+++ analysis/npm.go
@@ -1,4 +1,4 @@
-package main
+package analysis
 
 import (
 	"encoding/json"
@@ -30,10 +30,10 @@ func getNPMLatest(pkg string) string {
 	return details.DistTags.Latest
 }
 
-var NPMPackageManager = pkgManager{
-	image: "gcr.io/ossf-malware-analysis/node",
-	commandFmt: func(pkg, ver string) string {
+var NPMPackageManager = PkgManager{
+	Image: "gcr.io/ossf-malware-analysis/node",
+	CommandFmt: func(pkg, ver string) string {
 		return fmt.Sprintf("analyze.js %s@%s", pkg, ver)
 	},
-	getLatest: getNPMLatest,
+	GetLatest: getNPMLatest,
 }
