+++ analysis/ruby-runner.sh
@@ -1,16 +1,5 @@
 #!/bin/bash
-
-REGISTRY=gcr.io/ossf-malware-analysis
-
+topic=${OSSMALWARE_WORKER_TOPIC:-projects/ossf-malware-analysis/topics/workers}
 for pkg in $(cat rubygems.txt); do
-  rand=$(cat /dev/urandom | tr -dc 'a-z0-9' | head -c 10)
-  kubectl run rubygems-$rand --image=$REGISTRY/analysis --restart='Never' \
-      --privileged \
-      --labels=install=1,package_type=rubygems \
-      --annotations=package_name=$pkg \
-      --annotations=package_version=test \
-      --requests="cpu=250m" -- \
-      analyze \
-      --package="rubygems/$pkg" \
-      --upload="gs://ossf-malware-analysis-results-test/rubygems/$pkg"
+  gcloud pubsub topics publish "$topic" --attribute=name=$pkg,ecosystem=rubygems
 done
